Ensure we always have a default device

+++ app/src/main/java/com/alpriest/energystats/ui/flow/battery/BatteryPowerViewModel.kt
@@ -14,7 +14,7 @@ class BatteryPowerViewModel(
 
     init {
         calculator = BatteryCapacityCalculator(
-            capacityW = configManager.batteryCapacity.toInt(),
+            capacityW = configManager.batteryCapacity.toDouble().toInt(),
             configManager.minSOC
         )
     }
